refactor(ui): extract Combobox into a standalone component
